Add energy saver
Some plug-ins' async functionality can drain the battery pretty quickly.
Need ability to change settings depending on whether the machine is
running on battery or not.

Got the idea from NeoMake Readme and this seems to be the main culprit
so going to start there initially before rolling out across the rest of
the config.

However, the script from the Readme only runs once so if you start
NeoVim while plugged in and then disconnect, you'll still drain the
battery faster. Added a timer to account for this, runs every 15
seconds.