Add wincent/terminus (NeoVim)
The original card related to file changed externally warnings and having
to manually reload all the time with `:e! %`.

I remembered that Greg Hurrell (wincent) covered this in one of his
screencasts which lead me to his plug-in wincent/terminus.

This plug-in auto-reloads changed files, improves mouse support (another
open card that can now be merged into this one), and a few other nice
features.

The only drawback was the plug-in didn't handle changing the cursor
well, this feature was disabled.