Remove wincent/ferret (NeoVim)
When I added this it didn't feel right: I already had global search
set-up with Denite so it would only have been another step to do the
replace. Also, didn't like the way this plug-in completely overrode the
quickfix config.

Turns out it's really easy to add the replace step but I ended up
making quite a few optimisations along the way.

I decided to remove ferret in this commit and deal with the rest of the
changes in the next.